Added a DartClassReader (stripped from DataClassGenerator) as another stage to separate concerns. Still adds 'lineNumber' information to DartClass instances, but no textual content. Sits at the same level as the JSONReader.
Renamed 'line' and 'linePos' to 'lineNumber' in some cases to gradually remove ambiguity.
Renamed ClassField as DartClassProperty (more specific and aligned with other type names).
Added (temporarily?) an 'abstract' property to DartClass (to use if no classContent is available, as it's the case with the new reader)./
Added JSDoc to characterizeProject.
Started renaming of 'clazzes' to 'theClasses' and 'clazz' to 'theClass' or 'aClass'.